Use _pings_related to make sure pings database are created properly first
