+++ src/__tests__/helpers/_pings_related.test.ts
@@ -1,0 +1,4 @@
+import { pingsTest } from "./pings.parttest";
+
+// https://github.com/facebook/jest/issues/6194#issuecomment-419837314
+describe("pings", pingsTest);
